Исправил недочеты
